requirements.txt
Flask
Pillow
pypdf
pdf2docx
gunicorn